Move the `TypeCollector` trait to the `ir` module
This commit moves the `TypeCollector` trait out from the `codegen` module and
into its own submodule in `ir::type_collector`. Additionally, it puts the
various `TypeCollector` trait implementations next to the types that each
implementation is for.
